chore: rename utils path
